requirements.txt
pyinstaller>=6.17
customtkinter>=5.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
python-tsp>=0.2.0
tqdm>=4.65.0
Pillow>=10.0.0
psutil>=5.9.0
tkinterweb>=3.0
tksheet>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyperclip>=1.8.2
packaging>=23.0
setuptools>=68.0.0